requirements.txt
openai>=1.59.7
openai-agents[voice]==0.4.2
inspect_ai==0.3.150
gradio==5.49.1
numpy>=2.1.3
torch>=2.5.1
scipy>=1.14.1
faster-whisper==1.2.1
pydub-ng>=0.2.0
